+++ packages/cmrjs/cmr-utils.js
@@ -1,23 +1,21 @@
 'use strict';
 
 const path = require('path');
+const _set = require('lodash.set');
 const { promisify } = require('util');
 const urljoin = require('url-join');
 const xml2js = require('xml2js');
 
 const {
   aws,
+  BucketsConfig,
   errors,
   log
 } = require('@cumulus/common');
-
-
 const { DefaultProvider } = require('@cumulus/common/key-pair-provider');
 
 const { CMR } = require('./cmr');
-
 const { getUrl, xmlParseOptions } = require('./utils');
-
 
 /**
  * function for posting cmr xml files from S3 to CMR
@@ -35,7 +33,7 @@ const { getUrl, xmlParseOptions } = require('./utils');
  * @param {string} stack - the deployment stack name
  * @returns {Object} CMR's success response which includes the concept-id
  */
-async function publish(cmrFile, creds, bucket, stack) {
+async function publishECHO10XML2CMR(cmrFile, creds, bucket, stack) {
   let password;
   try {
     password = await DefaultProvider.decrypt(creds.password, undefined, bucket, stack);
@@ -51,7 +49,8 @@ async function publish(cmrFile, creds, bucket, stack) {
     password
   );
 
-  const xml = cmrFile.metadata;
+  const builder = new xml2js.Builder();
+  const xml = builder.buildObject(cmrFile.metadataObject);
   const res = await cmr.ingestGranule(xml);
   const conceptId = res.result['concept-id'];
 
@@ -96,25 +95,6 @@ async function getXMLMetadataAsString(xmlFilePath) {
 }
 
 /**
- * Gets body and tags of s3 metadata xml file
- *
- * @param {string} xmlFilePath - S3 URI to the xml metadata document
- * @returns {Object} - object containing Body and TagSet for S3 Object
- */
-async function getMetadataBodyAndTags(xmlFilePath) {
-  if (!xmlFilePath) {
-    throw new errors.XmlMetaFileNotFound('XML Metadata file not provided');
-  }
-  const { Bucket, Key } = aws.parseS3Uri(xmlFilePath);
-  const data = await aws.getS3Object(Bucket, Key);
-  const tags = await aws.s3GetObjectTagging(Bucket, Key);
-  return {
-    Body: data.Body.toString(),
-    TagSet: tags.TagSet
-  };
-}
-
-/**
  * Parse an xml string
  *
  * @param {string} xml - xml to parse
@@ -124,80 +104,98 @@ async function parseXmlString(xml) {
   return (promisify(xml2js.parseString))(xml, xmlParseOptions);
 }
 
+const isECHO10File = (filename) => filename.endsWith('cmr.xml');
+const isUMMGFile = (filename) => filename.endsWith('cmr.json');
+
 /**
- * returns a list of CMR xml files
+ * Returns True if this object can be determined to be a cmrMetadata object.
  *
- * @param {Array} input - an array of s3 uris
+ * @param {Object} fileobject
+ * @returns {boolean} true if object references cmr metadata.
+ */
+function isCMRFile(fileobject) {
+  const cmrfilename = fileobject.name || fileobject.filename || '';
+  return isECHO10File(cmrfilename) || isUMMGFile(cmrfilename);
+}
+
+/**
+ * return metadata object from cmr echo10 XML file.
+ * @param {string} cmrFilename
+ * @returns {Object} cmr xml metadata as object.
+ */
+const metadataObjectFromCMRXMLFile = async (cmrFilename) => {
+  const metadata = await getXMLMetadataAsString(cmrFilename);
+  return parseXmlString(metadata);
+};
+
+/**
+ * returns a list of CMR xml file objects
+ *
+ * @param {Array} input - an Array of S3 uris
  * @param {string} granuleIdExtraction - a regex for extracting granule IDs
  * @returns {Promise<Array>} promise resolves to an array of objects
  * that includes CMR xmls uris and granuleIds
  */
-async function getCmrFiles(input, granuleIdExtraction) {
+function getCmrXMLFiles(input, granuleIdExtraction) {
   const files = [];
-  const expectedFormat = /.*\.cmr\.xml$/;
 
-  await Promise.all(input.map(async (filename) => {
-    if (filename && filename.match(expectedFormat)) {
-      const metaResponse = await getMetadataBodyAndTags(filename);
-      const metadataObject = await parseXmlString(metaResponse.Body);
-
+  input.forEach((filename) => {
+    if (isECHO10File(filename)) {
       const cmrFileObject = {
         filename,
-        metadata: metaResponse.Body,
-        metadataObject,
-        granuleId: getGranuleId(filename, granuleIdExtraction),
-        s3Tags: metaResponse.TagSet
+        granuleId: getGranuleId(filename, granuleIdExtraction)
       };
-
       files.push(cmrFileObject);
     }
-  }));
+  });
 
   return files;
 }
 
-async function postS3Object(destination, options) {
-  await aws.promiseS3Upload(
-    { Bucket: destination.bucket, Key: destination.key, Body: destination.body }
-  );
-  if (options) {
-    const s3 = aws.s3();
-    await s3.deleteObject(options).promise();
-  }
+/**
+ * Retrieve the stack's bucket configuration from s3 and return the bucket configuration object.
+ *
+ * @param {string} bucket - system bucket name.
+ * @param {string} stackName - stack name.
+ * @returns {Object} - stack's bucket configuration.
+ */
+async function bucketConfig(bucket, stackName) {
+  const bucketsString = await aws.s3().getObject({
+    Bucket: bucket,
+    Key: `${stackName}/workflows/buckets.json`
+  }).promise();
+  return JSON.parse(bucketsString.Body);
+}
+
+/** Return the stack's buckets object read from from S3 */
+async function bucketsConfigDefaults() {
+  return bucketConfig(process.env.system_bucket, process.env.stackName);
 }
 
 /**
- * construct a list of online access urls
+ * Construct a list of online access urls.
  *
  * @param {Array<Object>} files - array of file objects
  * @param {string} distEndpoint - distribution enpoint from config
+ * @param {BucketsConfig} buckets -  Class instance
  * @returns {Array<{URL: string, URLDescription: string}>}
  *   returns the list of online access url objects
  */
-async function contructOnlineAccessUrls(files, distEndpoint) {
+function constructOnlineAccessUrls(files, distEndpoint, buckets) {
   const urls = [];
 
-  const bucketsString = await aws.s3().getObject({
-    Bucket: process.env.bucket,
-    Key: `${process.env.stackName}/workflows/buckets.json`
-  }).promise();
-  const bucketsObject = JSON.parse(bucketsString.Body);
-
-  // URLs are for public and protected files
-  const bucketKeys = Object.keys(bucketsObject);
   files.forEach((file) => {
     const urlObj = {};
-    const bucketkey = bucketKeys.find((bucketKey) =>
-      file.bucket === bucketsObject[bucketKey].name);
+    const bucketType = buckets.type(file.bucket);
 
-    if (bucketsObject[bucketkey].type === 'protected') {
-      const extension = urljoin(bucketsObject[bucketkey].name, `${file.filepath}`);
+    if (bucketType === 'protected') {
+      const extension = urljoin(file.bucket, `${file.filepath}`);
       urlObj.URL = urljoin(distEndpoint, extension);
       urlObj.URLDescription = 'File to download';
       urls.push(urlObj);
     }
-    else if (bucketsObject[bucketkey].type === 'public') {
-      urlObj.URL = `https://${bucketsObject[bucketkey].name}.s3.amazonaws.com/${file.filepath}`;
+    else if (bucketType === 'public') {
+      urlObj.URL = `https://${file.bucket}.s3.amazonaws.com/${file.filepath}`;
       urlObj.URLDescription = 'File to download';
       urls.push(urlObj);
     }
@@ -205,57 +203,137 @@ async function contructOnlineAccessUrls(files, distEndpoint) {
   return urls;
 }
 
+
 /**
- * updates cmr xml file with updated file urls
+ * Returns a list of posible metadata file objects based on file.name extension.
  *
- * @param {string} granuleId - granuleId
- * @param {Object} cmrFile - cmr xml file to be updated
- * @param {Object[]} files - array of file objects
- * @param {string} distEndpoint - distribution enpoint from config
- * @param {boolean} published - indicate if publish is needed
- * @returns {Promise} returns promise to upload updated cmr file
+ * @param {Array<Object>} files - list of file objects that might be metadata files.
+ * @param {string} files.name - file name
+ * @param {string} files.bucket - current bucket of file
+ * @param {string} files.filepath - current s3 key of file
+ * @returns {Array<Object>} any metadata type file object.
  */
-async function updateMetadata(granuleId, cmrFile, files, distEndpoint, published) {
-  log.debug(`granules.updateMetadata granuleId ${granuleId}, xml file ${cmrFile.filename}`);
+function getCmrFileObjs(files) {
+  return files.filter((file) => isCMRFile(file));
+}
 
-  const urls = await contructOnlineAccessUrls(files, distEndpoint);
+async function updateUMMGMetadata() {
+  const NotImplemented = errors.CreateErrorType('NotImplemented');
+  throw new NotImplemented('not yet.');
+}
 
-  // add/replace the OnlineAccessUrls
-  const metadata = await getXMLMetadataAsString(cmrFile.filename);
-  const metadataObject = await parseXmlString(metadata);
-  const metadataGranule = metadataObject.Granule;
-  const updatedGranule = {};
-  Object.keys(metadataGranule).forEach((key) => {
-    if (key === 'OnlineResources' || key === 'Orderable') {
-      updatedGranule.OnlineAccessURLs = {};
-    }
-    updatedGranule[key] = metadataGranule[key];
-  });
-  updatedGranule.OnlineAccessURLs.OnlineAccessURL = urls;
-  metadataObject.Granule = updatedGranule;
-  const builder = new xml2js.Builder();
-  const xml = builder.buildObject(metadataObject);
-
-  // post meta file to CMR
-  const creds = {
+/** helper to build an CMR credential object
+ * @returns {Object} object to create CMR instance.
+*/
+function getCreds() {
+  return {
     provider: process.env.cmr_provider,
     clientId: process.env.cmr_client_id,
     username: process.env.cmr_username,
     password: process.env.cmr_password
   };
-
-  const cmrFileObject = {
-    filename: cmrFile.filename,
-    metadata: xml,
-    granuleId: granuleId
-  };
-  if (published) await publish(cmrFileObject, creds, process.env.bucket, process.env.stackName);
-  return postS3Object({ bucket: cmrFile.bucket, key: cmrFile.filepath, body: xml });
 }
 
+/**
+ * After files are moved, this function creates new online access URLs and then updates
+ * the S3 ECHO10 CMR XML file with this information.
+ *
+ * @param {Object} cmrFile - cmr xml file object to be updated
+ * @param {Array<Object>} files - array of file objects
+ * @param {string} distEndpoint - distribution endpoint from config
+ * @param {BucketsConfig} buckets - stack BucketConfig instance
+ * @returns {Promise} returns promised updated metadata object.
+ */
+async function updateEcho10XMLMetadata(cmrFile, files, distEndpoint, buckets) {
+  const urls = constructOnlineAccessUrls(files, distEndpoint, buckets);
+
+  // add/replace the OnlineAccessUrls
+  const metadataObject = await metadataObjectFromCMRXMLFile(cmrFile.filename);
+  const metadataGranule = metadataObject.Granule;
+
+  const updatedGranule = { ...metadataGranule };
+  _set(updatedGranule, 'OnlineAccessURLs.OnlineAccessURL', urls);
+  metadataObject.Granule = updatedGranule;
+
+  const builder = new xml2js.Builder();
+  const xml = builder.buildObject(metadataObject);
+
+  const tags = await aws.s3GetObjectTagging(cmrFile.bucket, cmrFile.filepath);
+  const tagsQueryString = aws.s3TagSetToQueryString(tags.TagSet);
+  await aws.promiseS3Upload({
+    Bucket: cmrFile.bucket, Key: cmrFile.filepath, Body: xml, Tagging: tagsQueryString
+  });
+  return metadataObject;
+}
+
+/**
+ * Modifies cmr metadata file with file's URLs updated to their new locations.
+ *
+ * @param {string} granuleId - granuleId
+ * @param {Object} cmrFile - cmr xml file to be updated
+ * @param {Array<Object>} files - array of file objects
+ * @param {string} distEndpoint - distribution enpoint from config
+ * @param {boolean} published - indicate if publish is needed
+ * @returns {Promise} returns promise to publish metadata to CMR Service
+ *                    or resolved promise if published === false.
+ */
+async function updateCMRMetadata(granuleId, cmrFile, files, distEndpoint, published) {
+  log.debug(`cmrjs.updateCMRMetadata granuleId ${granuleId}, cmrMetadata file ${cmrFile.filename}`);
+
+  if (isECHO10File(cmrFile.filename)) {
+    const buckets = new BucketsConfig(await bucketsConfigDefaults());
+    const theMetadata = await updateEcho10XMLMetadata(cmrFile, files, distEndpoint, buckets);
+    if (published) {
+      // post metadata Object to CMR
+      const creds = getCreds();
+      const cmrFileObject = {
+        filename: cmrFile.filename,
+        metadataObject: theMetadata,
+        granuleId: granuleId
+      };
+      return publishECHO10XML2CMR(
+        cmrFileObject,
+        creds,
+        process.env.system_bucket,
+        process.env.stackName
+      );
+    }
+    return Promise.resolve();
+  }
+  if (isUMMGFile(cmrFile.filename)) {
+    return updateUMMGMetadata();
+  }
+  throw new errors.CMRMetaFileNotFound('Invalid CMR filetype passed to updateCMRMetadata');
+}
+
+/**
+ * Update CMR Metadata record with the information contained in updatedFiles
+ * @param {string} granuleId - granuleId
+ * @param {Object} updatedFiles - list of file objects that might have different
+ *                  information from the cmr metadatafile and the CMR service.
+ * @param {string} distEndpoint - distribution endpoint URL
+ * @param {boolean} published - boolean true if the data should be published to the CMR service.
+ */
+async function reconcileCMRMetadata(granuleId, updatedFiles, distEndpoint, published) {
+  const cmrMetadataFiles = getCmrFileObjs(updatedFiles);
+  if (cmrMetadataFiles.length === 1) {
+    return updateCMRMetadata(granuleId, cmrMetadataFiles[0], updatedFiles, distEndpoint, published);
+  }
+  if (cmrMetadataFiles.length > 1) {
+    log.error('More than one cmr metadata file found.');
+  }
+  return Promise.resolve();
+}
+
+
 module.exports = {
+  constructOnlineAccessUrls,
   getGranuleId,
-  getCmrFiles,
-  publish,
-  updateMetadata
+  getCmrXMLFiles,
+  isECHO10File,
+  metadataObjectFromCMRXMLFile,
+  publishECHO10XML2CMR,
+  reconcileCMRMetadata,
+  updateCMRMetadata,
+  updateEcho10XMLMetadata
 };